requirements.txt
click==8.1.8
dotenv==0.9.9
jinja2==3.1.5
openai==1.61.1
pandas==2.2.3
pre-commit==3.7.0
python-dateutil==2.9.0.post0
requests==2.32.3
